Add mikroorm config to allow using mikroorm CLI
Mikro-orm CLI used for running migrations. Note that the following typing strategy was used on the mikro-orm config: Parameters<typeof MikroORM.init>[0]. This allows us to type the config to the Parameters of the MikroORM.init function. Since parameters returns an array, we get the first item in the array.