Add a base server with a dummy entity (POST)
The base server connects to the database per the mikro-orm config file, and runs any migrations automatically. It then runs the create first post commands.

Note that the migration needs to first be created using the CLI: npx mikro-orm migration:create. That will create a migration file in the ./migrations folder which is then run via the `await orm.getMigrator().up();` command.

- An entity is required, so a dummy POST entity created using the CLASS decorator style (this is done since it plays nice with the type graphql dependency we will use later;
- main function used in the server to avoid any headaches with top-level async/await.
- simple dummy migration created usingnpx mikro-orm migration:create to create the table for POST based on the entity schema.