Create a test setup flow
To properly test code using integration tests, a test setup scaffolding is created that creates a new database for testing and creates the schema to initialize the database with the proper schema. After the new database is created and before we create the schema, the tables are dropped to clean the db completely every time the test setup is run. This is to avoid contamination of the test database and keep things properly isolated.

* testConn.ts -- module that returns the test connection initialized with test database arguments. Before the return, it creates a test database. If the db exists, it will throw an error, but that we are ignoring for now as it doesn't matter. TODO: programmatically check if db exists then create if not.
* runs testConn function and then drops the db tables and generates tables based on schema. This is done to ensure that db is clean
* gqlCall.ts -- abstracts call to graphql function for passing queries to resolvers.