Add error handling to register and login resolvers
Register and login both leverage the FieldError and UserResponse object types. The FieldError objectType returns any errors or null if no errors. The UserResponse returns FieldError as an array or a User (but never both). If there is an error, User will be null. If no errors, errors will be null. Added some validation if statements to register and login checking for minimum lengths for username and password. Register also checks if username already exists.