Comment out the "should return current user if logged in user test
This test is currently non-functioning and, given it was partially written, the watch server was throwing errors due to unused variables, etc. For now, just commenting it out will resolve any noise in the watch server console. This test will be addressed with a re-write of the test server logic using https://github.com/zapier/apollo-server-integration-testing given the limitations with the current apollo-test-server.