Add grapql code generation functionality
By default, when using mutations and queries with urql, there are no types associated with the queries/mutations. In order to get the types, install graphql code generator, which will allow us to generate the types based on the queries/mutations. The generator also generates a new, typed hook based on urql's useQuery and useMutation.