Fix issue with NavBar not showing logged in state
Urql caches mutations and queries which causes the logout state to be shown after a user logs in or registers. In reality, the NavBar should show the logged in state (show username + a logout button). To resolve, run cache update manually in the cacheExchange.

see https://formidable.com/open-source/urql/docs/graphcache/cache-updates/ for more info.