Avoid cache issue for logout whereby user looks to still be logged in
THe NavBar shows the username and a logout button if the user is logged in. Due to URQL caching, the NavBar remains with logged in state after a logout. To fix, update the me query upon logout to return NULL, which causes all listeners to update with the proper logged out state. In this case the NavBar correctly re-renders with the correct state.