Add SSR to login page.
After adding URQL SSR package, for queries to work, we need to add SSR (without enabling SSR) higher order component for the query to work. If we wanted SSR to be enabled, we would pass in the {ssr: true} object to the higher order component as the second argument. In this case, the higher order component is used so that the query works but we do not need to actually enable SSR.