Fix issue with redundant query for SSR pages
When pages are set to SSR, any query on a component that runs on the page will run on the nextjs server due to how SSR works. In this case NavBar runs the me query on the server, but since the server is not storing a cookie, the Me query will always return null. This can be fixed by passing the cookie to the server, but it's better to just avoid running teh me query on the server altogether by using the pause attribute.

* create an isServer util since we want to pause the query only when we're on the server but let it run if were on the client. The isServer function checks for the window object, which would only exist on the client. If typeof is undefined, then we're on the server.