Add more robust email validation
Email validation previously checked if an @ existed in the email. This is subpar so install the deep-email-validator package and set to check regex, typos, if email is disposable, and SMTP. Mx check was removed as that slows down the API and is not needed at this time. Later on, if needed, we can revisit.

* Add validateEmail util function
* Update validateRegister function to use validateEmail
* Update user resolver register function to await validateRegister as that was changed to an async function to work with deep-email-validator.