Add ForgotPassword functionality
ForgotPassword page simply allows the user to invoke the ForgotPassword mutation, which sends a tokenized email to the user so that she can change her password. There is no validation done on the account, so the email will only be sent if the email is associated to an account in the system.