Cleanup user test, add ChangePassword test (but skipped)
ChangePassword test is currently giving problems, so skipping it. The issue is that v4 function from uuid is mocked but is used by the entity in creating the user. Therefore, the mocked implementation causes issue. Using jest.spyOn is not an option since the function has configurable set to false.

Confidence for the user resolvers is high nonetheless, so this missing test will be addressed in the future.