Add fixtures and helper for tests.
Add fixtures for user and txn as well as a helper to create simple uuid. The fixtures will be used to populate the database. The simple uuid creator will allow us to differentiate simply the test data from any real data in the event that the populate script is accidentally used in a production database.