Ignore eslint rule for tests (no-explicit-any)
Tests should take into account best practices for typing where possible, but this should not be as strict as the rest of the codebase given that we want to optimize for test confidence and dev time. Where use of any is far simpler and does not reduce confidence or cause dev tech debt, it should be allowed.