refactor: add linting rule to ignore @typescript-eslint/explicit-module-boundary-types
application does not need a return type since it's obvious by the return statement. To fix, we would need to create an interface and it's just adding another thing to maintain without providing any benefits as the type system already shows what is returned. The linting error in this case can be safely ignored as it's not helping but would hinder with additional maintenance, albeit minor maintenance.