test: fix jest bug with spurious test failing
When running tests in parallel, jest was failing with the following error: Jest did not exit one second after the test run has completed. Apart from this, certain tests in transaction.test were also failing randomly. This is likely related and due to an issue with async functions not closing or not closing in time.

The fix is temporary as ideally jest should run in parallel for speed. THe command for jest was modified to jest --runInBand --detectOpenHandles --forceExit  which runs sequentially (runInBand), closes any open handles that have not closed (async issues) and forceExit if there is an issue rather than just hanging. Again, there is obviously an issue but this at least let's us run tests.