test: add comment to  commented out mock
Ideally, we should mock the isAuth function and just return next(), but you were getting an error related to next not being a function. This seems to be an issue wtih apollo-testing-integration not passing in next, but you're not sure.

See https://stackoverflow.com/questions/67768435/proper-way-to-test-type-graphql-middleware-with-jest for more info on the issue.

If this gets resolved, you will refactor the test with a mockImplementation that simply runs next()