refactor: createTransaction permitted only for logged in user
Previously, any userId could be passed into the resolver function to associate the created txn to a user. isAuth middleware simply checked that a user was logged in but not that the userId passed into the createTransaction function was the same as the logged in user.

To make things more secure, userId is no longer an argument for createTransaction. The associated user is simply the logged in user.

* update mutation
* fix "should create a txn successfully" test