feat: category entity migration and population
Create a migration for category and a category table population script,  which populates the category table with the CategoryName values from the enum in Category.ts. This script was to  facilitate things, but it's probably better for the population to occur in the migration SQL. THe population script checks if the entity has been populated. If it has, it will return. Otherwise, it will create the entries based on the enum.

- create population script
- create migration to create category entity
- run population script as part of application