fix: amount stored with cents rounded to decimal
The transaction  amount field, when stored with cents, was getting rounded to integer amounts due to the amount field being int4. Rather than convert the type to numeric with decimal precision, the best practice is to convert the amount to cents and store as cents (example: 100.99 => 10099) by multiplying by 100. To show the amount, we will divide by 100. The advantage to this approach is speed in any calculations.

https://stackoverflow.com/questions/15726535/postgresql-which-datatype-should-be-used-for-currency