test: refactor CreateTXN test to be more robust
Previously the test was merely checking the id in the db was the same as the newly created transaction. To make better, check all relevant fields (except createdAt, updatedAt,  and txnDate). This turned up missing mutation return for category, which was not being passed back upon creation, so that was updated as well.