feat: Add Category resolver to return all categories
This feature is used in the createTransaction page for the client. All categories will be obtained and  filtered based on the transaction type. FOr example, an income transaction type will filter categories to only be those that match "income". The same will happen for "expense" type transactions. Given there are only 19 categories and we don't anticipate adding many more, the simplest way to  handle this feature is to simply return all categories and then filter the categories client side.